test: composer-trigger suite stewardship + cross-surface parity
- Fold fileSearchIndexService tests into fileService.test.ts (helper has a single parent module; removes the fragment file)
- AgentChatComposer.test.tsx +3: mid-sentence slash splice, Enter fall-through on unmatched token (quality A1 regression), mid-sentence @file splice + attachment
- iOS: WorkComposerTriggerDetectorTests (14 pure-logic tests; suite green on simulator)
- Docs parity: composer-and-ui source map (composerTriggers + ChatCommandMenu rows), ade-code composer section, ios-companion composer section, file-watcher session-lane cache note
- CLI/TUI parity verified clean (headless daemon shares the action registry; TUI fall-through matches desktop)
